Separate Crossword pages and Tag pages, move shared logic into common.
