+++ applications/app/pages/CrosswordResultsHtmlPage.scala
@@ -7,24 +7,20 @@ import html.{HtmlPage, Styles}
 import model.ApplicationContext
 import play.api.mvc.RequestHeader
 import play.twirl.api.Html
-import services.TagPage
 import views.html.fragments._
 import views.html.fragments.commercial.pageSkin
 import views.html.fragments.page.body.{bodyTag, breakingNewsDiv, mainContent, skipToMainContent}
 import views.html.fragments.page.head.stylesheets.{criticalStyleInline, criticalStyleLink, styles}
 import views.html.fragments.page.head.{fixIEReferenceErrors, headTag, titleTag, weAreHiring}
 import views.html.fragments.page.{devTakeShot, htmlTag}
-import views.html.stacked
+import views.html.fragments.crosswords.{crosswordResultsPageBody, crosswordResultsPageHead}
+import model.CrosswordResultsPage
 
-object TagHtml {
+object CrosswordResultsHtmlPage extends HtmlPage[CrosswordResultsPage] {
 
   def allStyles(implicit applicationContext: ApplicationContext, request: RequestHeader): Styles = new Styles {
-    override def criticalCssLink: Html = stacked(
-      criticalStyleLink(FaciaCSSFile),
-      criticalStyleLink(InlineNavigationCSSFile))
-    override def criticalCssInline: Html = criticalStyleInline(
-      Html(common.Assets.css.head(Some("facia"))),
-      Html(common.Assets.css.inlineNavigation))
+    override def criticalCssLink: Html = criticalStyleLink(ContentCSSFile)
+    override def criticalCssInline: Html = criticalStyleInline(Html(common.Assets.css.head(Some("facia"))))
     override def linkCss: Html = stylesheetLink(s"stylesheets/$FaciaCSSFile.css")
     override def oldIECriticalCss: Html = stylesheetLink(s"stylesheets/old-ie.head.$FaciaCSSFile.css")
     override def oldIELinkCss: Html = stylesheetLink(s"stylesheets/old-ie.$ContentCSSFile.css")
@@ -32,18 +28,15 @@ object TagHtml {
     override def IE9CriticalCss: Html = stylesheetLink(s"stylesheets/ie9.$ContentCSSFile.css")
   }
 
-  def html(page: TagPage)(
-    headContent: Html = Html(""),
-    bodyContent: Html = Html("")
-  )(implicit request: RequestHeader, applicationContext: ApplicationContext): Html = {
-    implicit val p: TagPage = page
+  def html(page: CrosswordResultsPage)(implicit request: RequestHeader, applicationContext: ApplicationContext): Html = {
+    implicit val p: CrosswordResultsPage = page
 
     htmlTag(
       headTag(
         weAreHiring() when WeAreHiring.isSwitchedOn,
         titleTag(),
         metaData(),
-        headContent,
+        crosswordResultsPageHead(page),
         styles(allStyles),
         fixIEReferenceErrors(),
         inlineJSBlocking()
@@ -55,7 +48,8 @@ object TagHtml {
         guardianHeaderHtml(),
         mainContent(),
         breakingNewsDiv(),
-        bodyContent,
+//        CrosswordResultsPageCleaner(page, crosswordResultsPageBody(page)),
+        crosswordResultsPageBody(page),
         footer(),
         inlineJSNonBlocking(),
         analytics.base()
@@ -64,19 +58,4 @@ object TagHtml {
     )
   }
 
-}
-
-object TagHtmlPage extends HtmlPage[TagPage] {
-  def html(page: TagPage)(implicit request: RequestHeader, applicationContext: ApplicationContext): Html =
-    TagHtml.html(page)(
-      headContent = tagPageHead(page),
-      bodyContent = TagPageCleaner(page, tagPageBody(page))
-    )
-}
-
-object AllTagHtmlPage extends HtmlPage[TagPage] {
-  def html(page: TagPage)(implicit request: RequestHeader, applicationContext: ApplicationContext): Html =
-    TagHtml.html(page)(
-      bodyContent = all(page)
-    )
 }
