+++ applications/app/views/fragments/crosswords/crosswordResultsPageBody.scala.html
@@ -1,0 +1,43 @@
+@(crosswordResultsPage: model.CrosswordResultsPage)(implicit request: RequestHeader, context: model.ApplicationContext)
+@import common.{Edition, PagePaths}
+@import views.html.fragments.containers.facia_cards.{container, containerScaffold}
+@import views.html.fragments.pagination
+@import layout.Front
+
+@if(crosswordResultsPage.page.metadata.id == "help/accessibility-help") {
+    @containerScaffold("Preferences", "accesibility preferences") {
+        <div id="js-accessibility-preferences">Loading&hellip;</div>
+    }
+}
+
+
+<div class="l-side-margins">
+    <div class="u-cf index-page" data-link-name="Front | @request.path" role="main">
+
+        @defining(Front.makeFront(crosswordResultsPage, Edition(request)).containers) { containers =>
+            @containers.map { containerDefinition =>
+                @container(
+                    containerDefinition = containerDefinition,
+                    frontId = Some(crosswordResultsPage.idWithoutEdition)
+                )
+            }
+        }
+
+        @crosswordResultsPage.page.metadata.pagination.map { paginationInstance =>
+            <div class="fc-container__pagination">
+                @pagination(crosswordResultsPage.page.metadata.webTitle, paginationInstance, PagePaths.fromId(crosswordResultsPage.idWithoutEdition))
+            </div>
+        }
+
+        @fragments.trendingTopics(crosswordResultsPage.faciaTrails, crosswordResultsPage.page.metadata.id, None)
+
+        <div class="fc-container fc-container--commercial">@fragments.commercial.commercialComponent()</div>
+
+        @defining(crosswordResultsPage.makeLinkedData) { linkedData =>
+            <script data-schema="@{linkedData.`@type`}" type="application/ld+json">
+                @Html(model.meta.LinkedData.toJson(linkedData))
+            </script>
+        }
+
+    </div>
+</div>
