+++ applications/app/views/fragments/crosswords/crosswordResultsPageHead.scala.html
@@ -1,0 +1,13 @@
+@(crosswordResultsPage: model.CrosswordResultsPage)(implicit request: RequestHeader)
+
+@import common.LinkTo
+@import model.Tag
+
+@crosswordResultsPage.page match {
+    case tag: Tag => {
+        <meta name="twitter:url" content="@LinkTo(request.uri)" />
+        <meta property="og:title" content="@views.support.Title(crosswordResultsPage.page)" />
+    }
+    case _ => { }
+}
+
