+++ tag/app/views/fragments/tagPageBody.scala.html
@@ -6,6 +6,7 @@
 @import views.html.fragments.pagination
 @import views.support.Commercial.container.isFirstNonThrasherContainer
 @import views.support.RenderClasses
+@import layout.Front.makeFront
 
 @if(tagPage.page.metadata.id == "help/accessibility-help") {
     @containerScaffold("Preferences", "accesibility preferences") {
@@ -26,7 +27,7 @@
                     @fragments.guBand()
                 }
 
-                @defining(TagPage.makeFront(tagPage, Edition(request)).containers) { containers =>
+                @defining(makeFront(tagPage, Edition(request)).containers) { containers =>
                         @containers.map { containerDefinition =>
                             @container(containerDefinition,
                                 frontProperties = FrontProperties.fromBranding(Edition(request), branding),
@@ -45,7 +46,7 @@
 
                 <div class="fc-container fc-container--commercial">@fragments.commercial.commercialComponent()</div>
 
-                @defining(TagPage.makeLinkedData(tagPage)) { linkedData =>
+                @defining(tagPage.makeLinkedData) { linkedData =>
                     <script data-schema="@{linkedData.`@type`}" type="application/ld+json">
                         @Html(model.meta.LinkedData.toJson(linkedData))
                     </script>
@@ -59,7 +60,7 @@
         <div class="l-side-margins">
             <div class="u-cf index-page" data-link-name="Front | @request.path" role="main">
 
-                @defining(TagPage.makeFront(tagPage, Edition(request)).containers) { containers =>
+                @defining(makeFront(tagPage, Edition(request)).containers) { containers =>
                     @containers.map { containerDefinition =>
                         @container(
                             containerDefinition = containerDefinition,
@@ -78,7 +79,7 @@
 
                 <div class="fc-container fc-container--commercial">@fragments.commercial.commercialComponent()</div>
 
-                @defining(TagPage.makeLinkedData(tagPage)) { linkedData =>
+                @defining(tagPage.makeLinkedData) { linkedData =>
                     <script data-schema="@{linkedData.`@type`}" type="application/ld+json">
                         @Html(model.meta.LinkedData.toJson(linkedData))
                     </script>
